move things into `bindings` module
